// max-most-frontend/src/Interfaces/companyType.ts
export type Address = {
  id: string;
  first_name: string;
  middle_name?: string;
  last_name: string;
  description?: string;
  label?: string;
  street1: string;
  street2?: string;
  street3?: string;
  street4?: string;
  address1?: string;
  address2?: string;
  city: string;
  state: string;
  zip: string;
  country: string;
  is_default?: boolean;
  is_shipping?: boolean;
  is_billing?: boolean;
  is_residental?: boolean;
  phone?: string;
  phone2?: string;
  email?: string;
  email2?: string;
  email3?: string;
  email4?: string;
  type?: string;
  fax?: string;
  company?: string;
  phoneNumbers: string[];
  emails: string[];
  // emails?: string[]
};

export interface Contact {
  id: string;
  email?: string;
  is_billing: boolean;
  is_shipping: boolean;
  title?: string;
  website?: string;
  first_name?: string;
  middle_name?: string;
  last_name?: string;
  organization_id?: string;
  authorize_to_purchase?: boolean;
  billing_address_id?: string;
  billing_address: Address;
  billing_phone?: string;
  shipping_address: Address;
  shipping_address_id?: string;
  is_active?: boolean;
  do_not_call?: boolean;
  do_not_email?: boolean;
  do_not_mail?: boolean;
  do_not_text?: boolean;
  is_department?: boolean;
  office_phone?: string;
  private_note?: Note[];
  created?: string;
  updated?: string;
  user: UserData;
}
// ...
export interface Note {
  id: string;
  text: string;
  type: string;
  created: string;
  updated: string;
}

export type CompanyData = {
  billing_address: Address;
  id: string;
  name: string;
  address: Address;
  address_id: string;
  number: string;
  brand_id: string;
  billing_contact_id: string;
  shipping_contact_id: string;
  billing_contact: Contact;
  shipping_contact: Contact;
  external_id: string;
  notes: Note[];
  private_note: Note[];
  tax_exempt_id: string;
  is_individual: boolean;
  is_tax_exempt: boolean;
  is_active: boolean;
  created: string;
  updated: string;
  is_trash: boolean;
};